# floats: use f64 functions where available